pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[tool.poetry]
name = "CodeSumma"
version = "0.1.0"
description = "CodeSumma is an AI-powered code summarization tool that streamlines the development process by generating concise, token-limited, summaries of Python codebases."
authors = ["RyanMac <ryan@updoot.co>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
tiktoken = "^0.3.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
